Name merge stages in rev-specs
Borrow magic pathspec style format ':(magic):' for use in
revision spec naming of merge stages to avoid the
difficulty remembering the stage numbers, as described
in $gmane/239135.

The magic refspec is not a portable POSIX, Windows or classic Mac filename.

However it has the same potential name conflicts as the existing
stage number format with more accepting filesystems.

Prompted-by: Antoine Pelisse <apelisse@gmail.com>
Signed-off-by: Philip Oakley <philipoakley@iee.org>