[PATCH] Provide a release-notes man page / guide
Use the symlinked RelNote file as a list-block to
act as the core of the release-notes man page.

The mantainers work load should not be affected.

Signed-off-by: Philip Oakley <philipoakley@iee.org>
---

The included lines are inset 11 spaces by the man page
asciidoc formatting so some are more than 80 characters long.
No negative indent mechanism for the list block was found -
suggestions welcomed.

The use of the stalenotes section from git(1) was rejected.

The stalenotes, with over two hundred links, has greater than
2 digits allowed in the trailing link list on the man page,
so would miss the hundreds digit.

Also the html links for the "documentation for release x.y.z"
(at kerenel.org) are to html files that are not actually made
by the makefile.