doc: format-patch: don't use origin as a branch name
Historically (5 Nov 2005 v0.99.9-46-g28ffb89) the git-format-patch used
'origin' as the upstream branch name. That name is now used as the nominal
name for the upstream remote.

While 'origin' would be DWIMmed (do what I mean) to be that remote's
primary branch, do not assume the reader is ready for such magic.

Likewise, do not use 'origin/master' which may not be up to date with the
remote, nor reflect the reader's master branch. The patch series should be
relative to the reader's view of 'git show-branch HEAD master'.

Use the more modern 'master' as the reference branch name.

Signed-off-by: Philip Oakley <philipoakley@iee.org>