doc: gitrevisions: space out special ref names
Put the special revision names (FETCH_HEAD, etc.) on their own new
lines so that they become visible as a list to the reader, particularly
for HTML formatted man pages. Use the trailing '+' method for a simple
linebreak <BR> rather than a paragraph.

In addition, identify the commands that modify ORIG_HEAD, as many examples
use it.

Signed-off-by: Philip Oakley <philipoakley@iee.org>