engine.pl: Fix i18n -o option in msvc buildsystem generator [draft]
5e9637c (i18n: add infrastructure for translating Git with gettext,
2011-11-18) introduced an extra '-o' option into the make file.
This breaks the msvc build if NO_GETTEXT is not set. While this was fixed in xx , the MSYSGIT is still at

The key feature is to determine the structure not the settings.

Be more tolerant.

 engine.pl code for extracting the git.sln for msvc gui-IDE.

Add tests to remove these non linker options, in same vein as
74cf9bd (engine.pl: Fix a recent breakage of the buildsystem
generator, 2010-01-22).

Extra notes: If the .version file matches the last tag then
there is no message on the stderr, otherwise a
GIT_VERSION = xxx
is emited. Thus a 'clean' build would heva an empty errs.txt file
(which can then be script deleted by engine.pl).

The msvc-build script (msysgit) has other stuff soing on so let the
presence of the err.txt stand in place of a return flag.

The msvc-build script would need updating to add in a compile step
to confirm it compiles OK.
And then a run tests step to see that that works OK.

Signed-off-by: Philip Oakley <philipoakley@iee.org>