doc: clone: clarify that the command creates the new directory
The existing text can be (mis)read two ways. Either the user should
have prepared a new directory into which clone will place the repository.
Or, git-clone itself creates the repository.

Careful users will create a directory 'just in case', and be frustrated by the resulting two levels
of directory created.

The 'git clone' may be contrasted with 'git init' which does not create
a directory, but expects that the user already has an existing directory.

Make clear that it is the latter.

Signed-off-by: Philip Oakley <philipoakley@iee.org>