doc: commit: --fixup/--squash can take a commit revision
Be clearer that the --fixup/--squash options can take any of the
gitrevisions methods of specifying a commit, not just a 'hash'.

Signed-off-by: Philip Oakley <philipoakley@iee.org>
---
v1
It's not immediately obvious what different forms the <commit>
option can take. Spell out, and refer to the git revisions guide,
that any of the revision methods will work. On a side note, if one
looks at the glossary, a <commit> links to <commit object> which links
to <object> (and only tangentially "about" revisions) which then
says "uniquely identified by the SHA-1", so it is easy to think one
should use the sha1 here.  I only discovered this while following
up other parts of this series!