fix snprintf conditional for VS2015
VS2015 does have a proper snprintf definition. It is easier to
detect this version in the C pre-processor.

Note that when _MSVC_VER is undefined (i.e. another compiler) it
defaults to zero, so the overall conditional becomes true and
snprintf is defined as is required by the BOGUS flag.

Signed-off-by: Philip Oakley <philipoakley@iee.org>