engine.pl: cope with < > chars
WIP: This isn't sufficient as it's part of a multi-qiuoting compatibility
problem. Should really replace/remove the '\'' first, and then simply
handle the < > chars.

The '\'' was introduced by 39c015c
("Fixes for ancient versions of GNU make", 2006-02-18)

The < > was introduced by 7c6ef2f
("[PATCH] ARM optimized SHA1 implementation", 2005-09-20)

Not sure why it's a problem now.
(other VS2015 upgrades may be an influence)

Signed-off-by: Philip Oakley <philipoakley@iee.org>