WIP: create new Visual Studio generation file
This is rebased on top of remotes/dscho-git/vs2008
b7307d5 ("wip", 2016-12-23)

I split the old generator into the boiler plate and the App/Lib parts.

Next is to recombine them in the same manner as @dscho's VS2015 example
so that we have a common app/lib function, and use the Nuget libraries.

The merge was just that bit to big to handle all at once without a
clear head and a bit of space!

The other side issue is that the Nuget version needed for older XP
systems has gone from the microsoft site - need to find out what they
have done with it!

Signed-off-by: Philip Oakley <philipoakley@iee.org>