Use size_t instead of 'unsigned long' for data in memory
Currently the length of data which is stored in memory is stored
in "unsigned long" at many places in the code base.
This is OK when both "unsigned long" and size_t are 32 bits,
(and is OK when both are 64 bits).
On a 64 bit Windows system am "unsigned long" is 32 bit, and
that may be too short to measure the size of objects in memory,
a size_t is the natural choice.

Improve the code base in "small steps", as small as possible.
The smallest step seems to be much bigger than expected.
See this code-snippet from convert.c:
        const char *ret;
        unsigned long sz;
        void *data = read_blob_data_from_index(istate, path, &sz);
        ret = gather_convert_stats_ascii(data, sz);

The corrected version looks like this:
        const char *ret;
        size_t sz;
        void *data = read_blob_data_from_index(istate, path, &sz);
        ret = gather_convert_stats_ascii(data, sz);

However, when the Git code base is compiled with a compiler that
complains that "unsigned long" is different from size_t, we end
up in this huge patch, before the code base cleanly compiles.

And: there is more to be done in the zlib interface.

Reviewed-by: Johannes Schindelin <johannes.schindelin@gmx.de>
Signed-off-by: Torsten Bögershausen <tboegi@web.de>
---
<20190413151850.29037-1-tboegi@web.de>
[PATCH v3 1/1] Use size_t instead of 'unsigned long' for data in memory
Sat, 13 Apr 2019 17:18:50 +0200

applied using `git apply -3 "20190413-[PATCH v3 1_1] Use size_t instead of _unsigned long_ for data in memory-38765.txt"`
 from a thunderbird export converted to unix line endings - note need for the -3 3-way merg option.

This is the 3rd version of the patch.
 - Dscho contributed with a code-review (converted 2 more unsigned long)
 - Thomas Braun and Philip Oakley have done more work:
   https://github.com/tboegi/git/pull/1
   Those changes are not part of my patch series
 - Applying this patch on 'pu' gives 3 conflicts (blame.c, packfile.[ch])

 Those three conflicts were for "the repository" changes where unsigned long
was still in use