Sourcetrail: sub-dir option for project storage
The Sourcetrail project and associated files can be in a sub-dir.
Simply create the sub-dir in advance. Export the compile_commands.json
direct to the sub-dir (embedded paths are adjusted). Import to Sourcetrail
select the `..` as the include directory, saving the sourcetrail project
also in the sub-dir.

Careful readers will notice that the database files are not included here
and can be recreated/indexed by Sourcetrail.

Signed-off-by: Philip Oakley <philipoakley@iee.email>