blame: option to show only blamed commits --no-boundary
For large files the will be many lines which are not blamed with the
revision range or date period which clutter the output.

Add an option to suppress the display of boundary commits lines, not
just the object id (`-b` option).

This is non-functional at this stage. The next commit will add
functionality.

Signed-off-by: Philip Oakley <philipoakley@iee.email>