doc: fixup/squash: remove ellipsis marks, use <line> for clarify
Ellipsis marks fail to hint at the typoe or style of the missing content.

Tell the reader what is missing, for easier comprehension.

Signed-off-by: Philip Oakley <philipoakley@iee.email>
---

The fixup/squash process could probably benefit from its own section
as there are many places for user interaction with the process.
This is a minimal shift toward such an improvement.