blame: add option to show only blamed commits `--blame-only`
For large files there will be many lines which are not blamed with the
revision range or date period which clutter the output.

Add an option to suppress the display of boundary commits lines, not
just the object id (`-b` option).

Useful for those not yet familir, or not using, *nix tooling.

Signed-off-by: Philip Oakley <philipoakley@iee.email>